requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
Pillow>=8.3.2
pytesseract>=0.3.8
pandas>=1.3.0
openpyxl>=3.0.0
selenium>=4.0.0
webdriver-manager>=3.5.3
playwright>=1.40.0
playwright-stealth>=1.0.6
cryptography>=41.0.0
pyyaml>=6.0